[refdata,qt,agile] Move Book/BookStatus/RegulatoryBookType Qt to RefdataPlugin
No cross-component leakage: BookController, BookStatusController, and
RegulatoryBookTypeController are now constructed and menu-wired by
RefdataPlugin instead of TradingPlugin. BookStatus and
RegulatoryBookType are fully owned (construction, menu, logout) by
RefdataPlugin, contributed to a new Reference Data > Book Codes
submenu (moved out of Trading Codes after manual QA flagged that
placement as wrong for reference-data lookups). Book itself is also
owned by RefdataPlugin, but TradingPlugin still needs a pointer for
its "&Books" menu action and the Portfolio/Org Explorer composite
views -- resolved via a new non-owning RefdataPlugin::book_controller()
accessor, looked up through PluginRegistry + dynamic_cast at login
time (safe by construction: RefdataPlugin's load_order runs before
TradingPlugin's on login, and after it on logout).

Fixed a real cross-shared-library RTTI bug found along the way:
RefdataPlugin lacked ORES_QT_REFDATA_EXPORT, so dynamic_cast across
the ores.qt.trading/ores.qt.refdata .so boundary failed to load with
"undefined symbol: _ZTIN4ores2qt13RefdataPluginE" -- same bug class as
PR #1449's BookController export fix.

Manual QA scenario: all 7 steps PASS. Also filed a UX capture: QA
Validation Runner gives no unsaved-changes warning when the app is
closed with pending scenario results.

Unblocks Book codegen drift remediation, which resumes with
regenerate_regulatory_book_type.

Story-ID: 0E63013B-3B02-4968-A9EA-ED85C00EC7AF
Task-ID: EE3C93E9-7907-4E18-BB82-3737E0E778C7
