+++ doc/agile/versions/v0/sprint_23/sprint.org
@@ -48,8 +48,8 @@ For the definitions of the themes see [[id:A064D838-F127-4DD6-BB42-9A7902039AEE]
 | [[id:F90CDE0A-9813-407C-A145-30BA001FAAA8][Sprint 22 leftover cleanup]] | BACKLOG | 2026-07-11 | | Close out four standalone leftover tasks parked at sprint 22 close. |
 | [[id:DCC5853D-1B41-4EB0-8A85-07E2A8FABC07][Refdata entity NATS event registrar audit]] | BACKLOG | 2026-07-11 | | Wire NATS event registrars for the 26 refdata entities still missing one. |
 | [[id:EA22E78F-E4CB-4E76-B854-F85EB4E6AF48][Book data model cleanup]] | STARTED | 2026-07-08 | | Carried from sprint 22 (3/9 tasks done): classification flags, hierarchy documentation, legal entity/branch investigation, rates centre, ledger_ccy rename, allowed-list analysis. Blocked on Book codegen drift remediation for its remaining new-work tasks. |
-| [[id:0E63013B-3B02-4968-A9EA-ED85C00EC7AF][Move book to refdata]] | STARTED | 2026-07-11 | | Move Book/BookStatus/RegulatoryBookType Qt CRUD + plugin wiring from TradingPlugin to RefdataPlugin (no cross-component leakage). Portfolio's equivalent move is intentionally separate for parallel work. Blocks Book codegen drift remediation. |
-| [[id:F0DAFC8D-3C94-4F97-B546-D7AF7469A29E][Book codegen drift remediation]] | STARTED | 2026-07-11 | | Regenerate book/book_status/regulatory_book_type against current codegen templates, restore any silently-dropped SQL validations, end-to-end test, then migrate book to composite-entity versioning and regen/test again. Blocked on Move book to refdata. |
+| [[id:0E63013B-3B02-4968-A9EA-ED85C00EC7AF][Move book to refdata]] | DONE | 2026-07-11 | 2026-07-11 | Moved Book/BookStatus/RegulatoryBookType Qt CRUD + plugin wiring from TradingPlugin to RefdataPlugin (no cross-component leakage), incl. a new Reference Data > Book Codes submenu. Portfolio's equivalent move is intentionally separate for parallel work. |
+| [[id:F0DAFC8D-3C94-4F97-B546-D7AF7469A29E][Book codegen drift remediation]] | STARTED | 2026-07-11 | | Regenerate book/book_status/regulatory_book_type against current codegen templates, restore any silently-dropped SQL validations, end-to-end test, then migrate book to composite-entity versioning and regen/test again. Unblocked, resuming. |
 | [[id:FA93094D-718E-42B0-A41A-434051CC6275][Resolve codegen model unification blockers]] | STARTED | 2026-06-27 | | Carried from sprint 22 (8/15 tasks done): B4 (collapse refdata/refdata-cpp), content migration, and table-model retirement remain. |
 | [[id:A083FE26-521D-4046-9E58-AFB901B28809][Codegen developer experience improvements]] | BACKLOG | 2026-06-25 | | Carried from sprint 22: downgrade address/type mismatch to skip in --component mode; add --entity filter; add entity group/tag system; literate template commentary. |
 | [[id:D6C90C13-555C-471F-B494-CAF6A7902519][Market data cleanup: retire dead duplicate tables and preserve hand-written overrides]] | BACKLOG | 2026-07-02 | | Carried from sprint 22: drop legacy marketdata tables (superseded, zero C++ references); remove stale validation_ignore.txt entries; protect market_observation's hand-written overrides from codegen regeneration. |
